chore(agent-auth): clear post-merge repo-shape red (dupe providers file, raw label, oversized SettingsScreen)
